+++ rmk/src/ble/esp/mod.rs
@@ -33,7 +33,7 @@ use rmk_config::StorageConfig;
 /// * `input_pins` - input gpio pins
 /// * `output_pins` - output gpio pins
 /// * `keyboard_config` - other configurations of the keyboard, check [RmkConfig] struct for details
-/// * `spwaner` - embassy task spwaner, used to spawn nrf_softdevice background task
+/// * `spawner` - embassy task spawner, used to spawn nrf_softdevice background task
 pub(crate) async fn initialize_esp_ble_keyboard_with_config_and_run<
     M: MatrixTrait,
     Out: OutputPin,
@@ -57,9 +57,7 @@ pub(crate) async fn initialize_esp_ble_keyboard_with_config_and_run<
     )
     .await;
 
-    let keymap = RefCell::new(
-        KeyMap::<ROW, COL, NUM_LAYER>::new_from_storage(default_keymap, Some(&mut storage)).await,
-    );
+    let keymap = RefCell::new(KeyMap::new_from_storage(default_keymap, Some(&mut storage)).await);
 
     let keyboard_report_sender = keyboard_report_channel.sender();
     let keyboard_report_receiver = keyboard_report_channel.receiver();
@@ -112,7 +110,7 @@ pub(crate) async fn initialize_esp_ble_keyboard_with_config_and_run<
         };
         let via_fut = vial_task(&mut via_rw, &mut vial_service);
         let matrix_fut = matrix.scan();
-        let storage_fut = storage.run::<ROW, COL, NUM_LAYER>();
+        let storage_fut = storage.run();
         pin_mut!(storage_fut);
         pin_mut!(via_fut);
         pin_mut!(keyboard_fut);