+++ rmk/src/keyboard.rs
@@ -78,6 +78,8 @@ pub(crate) async fn communication_task<'a, W: HidWriterWrapper, W2: HidWriterWra
     keybooard_hid_writer: &mut W,
     other_hid_writer: &mut W2,
 ) {
+    // This delay is necessary otherwise this task will stuck at the first send when the USB is suspended
+    Timer::after_secs(2).await;
     loop {
         match receiver.receive().await {
             KeyboardReportMessage::KeyboardReport(report) => {
@@ -354,7 +356,7 @@ impl<'a, const ROW: usize, const COL: usize, const NUM_LAYER: usize>
                 self.process_action_keycode(key, key_event).await;
                 self.update_osm(key_event).await;
                 self.update_osl(key_event);
-            },
+            }
             Action::LayerOn(layer_num) => self.process_action_layer_switch(layer_num, key_event),
             Action::LayerOff(layer_num) => {
                 // Turn off a layer temporarily when the key is pressed